[spv-out] Move Writer type to back/spv/mod.rs.
The `impl` blocks for `Writer` and its associated types remain in spv/writer.rs.

There should be no substantive code changes in this commit, only moving code
around and adjusting `use` declarations to match.
